Working on the "activePI" and "calculateTurnover" functions in a separate sandbox.
In Retention Exploratory Data Analysis:

Modifying the "Exploratory Data Analysis" to follow an outline where I first evaluate the HR retention data separtely, and then repeat the evaluation as informed by the proposal data.

Adding two visualizations to look at hire by quarter and year as well as by week.  These all need help with the labels, and the calculation of "turnover" frame will probably change based on my sandbox.

Investigating the "tendrils" by plotting them as lines.  Seeing that they are sudden drops in the head count. A closer looks shows that they aren't aligning with termination and hires, which triggered the sandbox work into the turnover calculations.

Experimented with graphics that show durations (or intervals) based on the proposal data.   I calculated an "effective hire" which is an effort to determine which "hire" date to use in my turnover calculations.