Wrote a third version of calculateTurnover, "part 3", and a sandbox to check it in.
In this third version, I use deltaHeadCount to calculate the daily head count (as delta.cum, as a default starting from the earliest period that I have.)  Then I aggregate to the correct period with the mean daily delta.cum and call this new column the headcount_mean for that period.

However, when I double-check it, the numbers especially for weeks seem time-shifted.

I am guessing that's either an "isoweek" problem, or probably more accurate how "isoweek" calculates my initial head count, or a shift in which period is aggregated and how it is merged back in.

This needs some additional double-checking.