Made adjustments in deltaHeadCount to accommodate the isoweek in PART 3, discovered, tested, and checked them in PART 3 CHECK, and copied them over to Turnover Functions.
my periodEnd column was using the default of week_start = 7 (Sunday) and I had to set that to 1 (Monday) in case where "calendar" argument was "week".

Somewhat similarly, I was taking the calendar year instead of the isoyear when aggregating, meaning dates like Jan 1st were often in the isoyear of the previous year.  These erroneous aggregations were then lost on merging.