The problem I'm trying to solve is the legend for the display of multiple lines.
Previous this legend was only on the "headcount" or "cumulative" graphic.  Then, what happens if I don't display it?

So I have been working on showing these on the right, with a complicated logical "ifelse" flow for managing legends on the right based on the number of lines.

It's too complicated and not working.  But I'll commit what I have before I make additional toggles.