Moved variables describing the target population (missing, single, multiple) earlier in the report.
Fiddled with words (purpose, objectives.)

Turned off a section with "eval = FALSE".

Wrote an executive summary concluding that this dataset is unattractive for using further.

Added a sandbox where I experimented with layout and drawing lines over all of the tile plots in a layout.  Generally, it's too fiddly to make it work well.