Lots of changes in Workforce Journey EDA (single records).
Flow was changed so that the tables appear at the end instead of the beginning.

A couple of barplot graphics were rotated.

A graphic describing row count by job title was added.

Words were modified to reflect the new query.  Some lines and paragraphs were deleted.

The tables with EMPLID examples of too many entries or exits were set  to "eval = FALSE" so they don't display.