Various thoughts and efforts in figuring out headcount for concurrent jobs.
Highlights include:

Creation of timeline graphics with jitter for multiple actions on the same date.

piSegments is a function that returns a list of entry and exit dates per boundary type

plotMap adds plot values (like color, size, and shape type) to actionFrame.

experimenting with using piSegments to draw lines aligning with the segments.

There's some good stuff in here, but it's not all the way there and is still somewhat muddled and confused.

Next I will work on iterating through the graphics and extractions, and I still need a good validation check