Created a "universityBoundaries" function which puts the entry/exit dates as pairs that are in order.
Then, an entry or exit is considered a university-wide entry or exit (as opposed to simply applying to a concurrent job) according to a logical comparison with other entries and exits.

Created a "plotJourney" which maps the actions for an individual employee on a timeline.

Created a function to extract the entry and exit dates per EMPLID.  This effort is abandoned and replaced with "universityBoundaries."

Created the function "assignBoundaries" which accepts a data frame of all unique actions and action reasons and assigns boundaries (entry/exit), types (primary, break, leave), and plot values based on the actions and action reasons.

Lots of work and thinking and iterations and half-starts in the Concurrent Jobs EDA sandbox.