Started to create a new "regime" for pre-2013 students, or for people who were re-hired after 2013 so I could ignore all pre-2013 information.
It worked great for many of my 20 random checks.

However, it failed a check where the gap was bridged by a second EMPL_RCD for a few of the 20 random checks.

It also identified a possibility for a similar regime where there was a large gap in the activity and then the EMPLID was re-hired just before 2013.

Continued to work on the sandbox, trying to make  a modification where there are max of two EMPL_RCD's and three or more entries.