Developed a sandbox for the "brute force" method, where I calculate the deltaHeadCount for a single employee at a time and I force the head count (delta cumulative) back to 1 if it exceeds 1 or back to zero if it goes below zero.
For ease of manipulating the functions, I copied the functions over into a "Brute Force Functions" file.  I haven't made any changes to these functions yet.

This approach looks promising and simplifies the tangling of multiple concurrent jobs with different kinds of overlapping breaks and leaves.

Next step is modify the functions and plots to show it, creating a function to extract the force-fit start and stop, and then adding in transfers for departments and titles.  Time between pay raises and pay grades is a nice next step.