Created an "extractUniversityBoundaries" function.
This takes each employee one at a time, force-fits a maximum of 1 or minimum of 0 cumulative head count for that one person, and extracts the dates of change.

These are the dates of a new  "university" boundary applicable to employees with concurrent jobs.

Next, I'd like to incorporate these dates into my current functions like "calculateMetrics."