Nearly achieved "force fit" headcount approximation.
Modified three functions to accommodate the new column, "univ_boundary", derived from a "force fit" estimation.  These functions are "deltaHeadCount_univ", "calculateMetrics_univ", and "plotMetrics_univ."  They differ in that they use the column "boundary_univ" and the values "stop" or "start" instead of "boundary" column and "entry" or "exit".

Two problems currently:  (1) The headcount is too high.  It is probably due to merging on a date that will multiply rows if multiple actions happen on a date.  (2)  Some of these employees stop having any new data but the last record isn't an official termination.   This would inflate the headcount.